♻ Make small refactoring global
