--- src/lib/stores/service.ts
@@ -1,5 +1,0 @@
-export function installService() {
-  var script = document.createElement("script");
-  script.textContent = "console.log(window);";
-  (document.head || document.documentElement).appendChild(script);
-}
